backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
pydantic==2.12.5
pydantic-settings==2.13.1
motor==3.7.1
httpx==0.28.1
PyJWT==2.10.1
python-dotenv==1.0.1
